tests/wpt/tests/css/css-text/parsing/word-spacing-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing word-spacing with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-word-spacing">
<meta name="assert" content="word-spacing supports only the grammar 'normal | <length-percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("word-spacing", "auto");

test_invalid_value("word-spacing", "normal 10px");
test_invalid_value("word-spacing", "10% 10px");
</script>
</body>
</html>
